refactor: Create folders to group functions
