+++ src/lib/reducers/content.js
