+++ src/lib/store/index.js
@@ -1,5 +1,5 @@
 import { createStore } from 'redux'
-import reducers from './reducers'
+import reducers from '../reducers'
 
 export default () => {
   const store = createStore(reducers)
